backend-fastapi/requirements.txt
fastapi
uvicorn[standard]
redis
websockets
PyJWT